+++ api/registry.go
@@ -4,6 +4,7 @@ import (
 	"encoding/json"
 	"fmt"
 	"net/http"
+	"time"
 
 	"github.com/wercker/wercker/util"
 )
@@ -18,15 +19,20 @@ type StepRegistry interface {
 
 // WerckerStepRegistry implements the StepRegistry interface to handle
 type WerckerStepRegistry struct {
-	baseURL string
+	baseURL   string
 	authToken string
+	logger    *util.LogEntry
 }
 
 // NewWerckerStepRegistry creates a new instance of NewWerckerStepRegistry
 func NewWerckerStepRegistry(baseURL, authToken string) StepRegistry {
+	logger := util.RootLogger().WithFields(util.LogFields{
+		"Logger": "Registry",
+	})
 	return &WerckerStepRegistry{
-		baseURL: baseURL,
+		baseURL:   baseURL,
 		authToken: authToken,
+		logger:    logger,
 	}
 }
 
@@ -34,7 +40,7 @@ func NewWerckerStepRegistry(baseURL, authToken string) StepRegistry {
 func (r *WerckerStepRegistry) GetStepVersion(owner, name, version string) (*APIStepVersion, error) {
 	url := fmt.Sprintf("%s/api/steps/%s/%s/%s", r.baseURL, owner, name, version)
 
-	resp, err := util.Get(url, r.authToken)
+	resp, err := r.getWithRetry(url, r.authToken)
 	if err != nil {
 		return nil, err
 	}
@@ -68,5 +74,26 @@ func (r *WerckerStepRegistry) GetStepVersion(owner, name, version string) (*APIS
 }
 
 func (r *WerckerStepRegistry) GetTarball(tarballURL string) (*http.Response, error) {
-	return util.Get(tarballURL, r.authToken)
+	return r.getWithRetry(tarballURL, r.authToken)
+}
+
+func (r *WerckerStepRegistry) getWithRetry(url string, authToken string) (*http.Response, error) {
+	var resp *http.Response
+	var err error
+	try := 0
+	maxTries := 3
+	for try < maxTries {
+		if try != 0 {
+			r.logger.Infof("Retrying step url %s %d", url, try)
+		}
+
+		resp, err = util.Get(url, authToken)
+		if err == nil {
+			break
+		}
+
+		time.Sleep(time.Second * 1)
+		try++
+	}
+	return resp, err
 }